requirements.txt
pandas
streamlit==0.76.0
Pillow
fake-useragent